Feature/message queue (#88)
* Move all writer files to writer folder
* Remove test for delete semaphores
* Backup
* Link libraries to pthread on windows
* Create thread, start and stop it
* Create queue mechanism
* Define message struct to cary message data and content
* Use new struct message
* Add custom data deleter to qnode_delete
* Writer writer thread function
* Update real's main.c to start 5 threads
* Fix typo
* Fix typo
* Add doc
* Minor style and warning fix + remove volatile keyword (useless in this case)
* Update changelog
* Add more thread on real test application
* Add extra swap of the queue to avoid remaining log message on closing
* Fix disabling subsystem when parent is disabled
* Adapt lock free queue to linux systems
* Adapt real application to linux
* Fix mistake avoiding build
* Fix mistake avoiding build
* Fix validation script
* Fix validation errors
* Fix style errors
* Fix windows build errors
* Add pthread for MSVC 12
* Remove old 'real' folder
* Improve real tests
* Install pthread on Windows
* Install pthread on windows

+++ include/yall/call.h
@@ -38,7 +38,7 @@ struct yall_call_data_line {
 };
 
 typedef struct yall_call_data {
-	uint16_t message_size;
+	size_t message_size;
 	char *header;
 	struct yall_call_data_line *lines;
 } yall_call_data;
